+++ src/app/modules/variable_pay/entities/vp-scoring-criterion.entity.ts
@@ -17,7 +17,7 @@ export class VpScoringCriterion extends BaseModel {
   @ManyToOne(() => VpCriteria, { onDelete: 'CASCADE' })
   vpCriteria: VpCriteria;
 
-  @Column({ type: 'decimal', precision: 10, scale: 2 })
+  @Column({ type: 'decimal', precision: 16, scale: 2 })
   weight: number;
 
   @Column({ type: 'uuid' })
